fix: explore section params code lines fixed and UI styles improved
