+++ app/experiment/explore/page.tsx
@@ -6,24 +6,24 @@ import InterviewSection from "@/app/experiment/explore/components/InterviewSecti
 import HackathonsSection from "@/app/experiment/explore/components/HackathonsSection";
 import Footer from "@/app/experiment/explore/components/Footer";
 import SiteHeader from "@/components/ui/site-header";
-
-import "@/app/experiment/explore/style/explore.css";
 import SiteFooter from "@/components/ui/site-footer";
 
+import "@/app/experiment/explore/style/explore.css";
+
 const Index = () => {
-  return (
-    <div className="min-h-screen bg-background">
-      <SiteHeader />
-      <ExploreHero />
-      <ExploreNavigation />
-      <SkillProgramsSection />
-      <PathsSection />
-      <InterviewSection />
-      <HackathonsSection />
-      <Footer />
-      <SiteFooter />
-    </div>
-  );
+    return (
+        <div className="min-h-screen bg-background">
+            <SiteHeader />
+            <ExploreHero />
+            <ExploreNavigation />
+            <SkillProgramsSection />
+            <PathsSection />
+            <InterviewSection />
+            <HackathonsSection />
+            <Footer />
+            <SiteFooter />
+        </div>
+    );
 };
 
 export default Index;
